Fix building with opus in non-standard prefix.
This moves the opus parts of JackNetTool into JackNetOpus so opus
headers are not undeliberately included from JackNetDriver (which is
compiled without opus).
All opus includes has got the opus/ prefix removed since it is added by
the pkg-config opus file.
Also the netone driver has to be compiled with opus.
Lastly JackNetTool.h needed a header guard.
