+++ common/netjack.h
@@ -33,8 +33,8 @@
 #endif
 
 #if HAVE_OPUS
-#include <opus/opus.h>
-#include <opus/opus_custom.h>
+#include <opus.h>
+#include <opus_custom.h>
 #endif
 
 #ifdef __cplusplus
